VisualizerApplication:
Created new class VisualizerApplication, which moves the thread start up functionality from main to this class.

This class will be threaded to consistently call on the Java Application thread for a user input version. while also extracting data from the application class.

The purpose of this class will be to allow more unique data extraction which would be abstracted in the VisualizerApplication class. The class will in the future interpret input to the VisualizerMediaPlayerHolder class through varying methods input (E.g. ActionEvents or Scanner String input). [General idea: the class will have an argument which allows a commandline argument to run string input through a terminal in pair with ActionEvents]